Add cloudpathlib storages
