+++ Makefile
@@ -14,23 +14,29 @@ check_release:
 	python setup.py sdist bdist_wheel
 	twine check dist/* --verbose
 
-clean:
-	find . -name "*.pyc" -exec rm -rf {} \;
-	find . -name "*.py,cover" -exec rm -rf {} \;
+clean_dev:
 	find . -name "*.orig" -exec rm -rf {} \;
 	find . -name "__pycache__" -exec rm -rf {} \;
-	rm -rf build/
-	rm -rf builddocs/
 	rm -rf dist/
 	rm -rf src/faker-file.egg-info/
 	rm -rf src/faker-file/var/
 	rm -rf .cache/
 	rm -rf .idea/
-	rm -rf htmlcov/
-	rm -rf .coverage
-	rm -rf .pytest_cache/
 	rm -rf .mypy_cache/
 	rm -rf .ruff_cache/
+
+clean_docs:
+	rm -rf build/
+	rm -rf builddocs/
+
+clean_test:
+	find . -name "*.pyc" -exec rm -rf {} \;
+	find . -name "*.py,cover" -exec rm -rf {} \;
+	rm -rf .coverage
+	rm -rf .pytest_cache/
+	rm -rf htmlcov/
+
+clean: clean_dev clean_docs clean_dev
 
 compile_requirements:
 	echo "common.in"
